+++ dev/io.openliberty.transaction.remote_fat/.settings/bndtools.core.prefs
