+++ dev/io.openliberty.transaction.remote_fat/fat/src/tests/EJBTest.java
@@ -1,0 +1,7 @@
+package tests;
+
+import componenttest.topology.utils.FATServletClient;
+
+public class EJBTest extends FATServletClient {
+
+}
